BI-6581 renaming package to retry
