+++ src/main/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/configuration/ApplicationConfig.java
@@ -4,7 +4,7 @@ import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.client.RestTemplate;
-import uk.gov.companieshouse.chipsrestinterfacesconsumer.producer.KafkaRestClient;
+import uk.gov.companieshouse.chipsrestinterfacesconsumer.retry.KafkaRestClient;
 
 @Configuration
 public class ApplicationConfig {
