+++ src/main/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/retry/AvroSerializer.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.chipsrestinterfacesconsumer.producer;
+package uk.gov.companieshouse.chipsrestinterfacesconsumer.retry;
 
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericDatumWriter;
