+++ src/main/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/retry/MessageProducer.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.chipsrestinterfacesconsumer.producer;
+package uk.gov.companieshouse.chipsrestinterfacesconsumer.retry;
 
 import uk.gov.companieshouse.chips.ChipsRestInterfacesSend;
 import uk.gov.companieshouse.service.ServiceException;
