+++ src/main/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/retry/MessageProducerImpl.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.chipsrestinterfacesconsumer.producer;
+package uk.gov.companieshouse.chipsrestinterfacesconsumer.retry;
 
 import org.apache.avro.Schema;
 import org.apache.kafka.clients.producer.RecordMetadata;
