+++ src/test/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/retry/RetryDelayTest.java
@@ -1,0 +1,9 @@
+package uk.gov.companieshouse.chipsrestinterfacesconsumer.retry;
+
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.junit.jupiter.MockitoExtension;
+
+@ExtendWith(MockitoExtension.class)
+public class RetryDelayTest {
+
+}
