+++ src/test/java/uk/gov/companieshouse/chipsrestinterfacesconsumer/service/impl/MessageProcessorServiceImplTest.java
@@ -9,7 +9,7 @@ import org.springframework.web.client.RestClientException;
 import uk.gov.companieshouse.chips.ChipsRestInterfacesSend;
 import uk.gov.companieshouse.chipsrestinterfacesconsumer.client.ChipsRestClient;
 import uk.gov.companieshouse.chipsrestinterfacesconsumer.common.ApplicationLogger;
-import uk.gov.companieshouse.chipsrestinterfacesconsumer.producer.MessageProducer;
+import uk.gov.companieshouse.chipsrestinterfacesconsumer.retry.MessageProducer;
 import uk.gov.companieshouse.service.ServiceException;
 
 import static org.mockito.ArgumentMatchers.eq;
